Allows project switching and removal
Users can click on a project in the nav bar to switch the active
project to the clicked project. Deleting a currently active project
will set the active project to "Inbox". Each project now has a hidden
indicator of whether it's the currently active project or not to
facilitate the above feature.